requirements.txt
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
expiringdict==1.2.2
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
proxy-tools==0.1.0
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
PyGObject==3.42.2
pyparsing==3.0.9
pywebview==4.0.1
Werkzeug==2.2.2
whichcraft==0.6.1
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
